+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 ### Changelog
+[2021-04-29] add wx.request interceptor, preparing fetch interceptor
 [2021-03-29] init xhr-response-mock, add XMLHttpRequest interceptor
 
 ### Features
+[2021-04-29]
+1. support wx.request response mock
+
 [2021-03-29]
 1. add XMLHttpRequest interceptor
 2. support GET, POST, PUT, PATCH and DELETE request mock